Update README.md
I thought github did this automatically, but it doesn't. Hopefully adding two spaces to the end of lines will force github to make a newline like reddit.

Also fixed youtube link.

To be honest, I'm not entirely sure if I like this change. It adds redundant spaces, and kinda looks weird.
Up to you to merge this or not - I recommend thinking about it and comparing the two before you decide.