Add melee mode to navbot
This doesn't just add the melee mode variable, it also adds a full melee AI that will kick in whenever a bot is using a melee weapon. This could easily be used to make bots counter fist of steal heavies for example, by simply making the bot switch to melee in getBestSlot.